fix: apply Android CJK font fallback to typography
