+++ apps/mobile/src/components/Typography.tsx
@@ -2,7 +2,9 @@
 import React from 'react';
 import {
   Platform,
+  StyleProp,
   Text as RNText,
+  TextStyle,
   TextInput as RNTextInput,
 } from 'react-native';
 import {
@@ -12,6 +14,10 @@ import {
 import AnimateableTextImpl from 'react-native-animateable-text';
 import { Text as RNEUITextImpl } from '@rneui/base';
 import Animated from 'react-native-reanimated';
+import {
+  containsCJKText,
+  sanitizeAndroidCJKFontStyle,
+} from './textFontFallback';
 
 // Since createAnimatedComponent return type is ComponentClass that has the props of the argument,
 // but not things like NativeMethods, etc. we need to add them manually by extending the type.
@@ -27,8 +33,21 @@ function withDefaults<C extends React.ComponentType<any>>(
   type Ref = React.ComponentRef<C>;
 
   const Component = (props: Props) => {
+    const AnyWrappedComponent = WrappedComponent as React.ComponentType<any>;
     const mergedProps = { ...defaultProps, ...props } as Props;
-    return <WrappedComponent {...mergedProps} />;
+    const textProps = mergedProps as Props & {
+      children?: React.ReactNode;
+      style?: StyleProp<TextStyle>;
+    };
+    return (
+      <AnyWrappedComponent
+        {...mergedProps}
+        style={sanitizeAndroidCJKFontStyle(
+          textProps.style,
+          containsCJKText(textProps.children),
+        )}
+      />
+    );
   };
 
   Component.displayName =
